requirements.txt
streamlit
routeros_api
google-generativeai